pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "VictoorLDS", email = "vlds221005@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "setuptools>=68.1.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "drf-spectacular>=0.26.4",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

